fix(input): Ctrl+Space系统切换时上屏编码内容
Ctrl+Space切换中英文时，系统通过Compartment回调触发，之前C++端
直接清空组合文本后异步通知Go，导致待输入编码丢失。

新增CMD_SYSTEM_MODE_SWITCH同步命令，让Go端在切换前检查
CommitOnSwitch配置并返回待上屏文本，C++端根据返回值决定
CommitText（上屏）或EndComposition（清空）。

同时修复KeyEventSink.cpp中循环变量有符号/无符号比较警告。

+++ wind_tsf/src/KeyEventSink.cpp
@@ -1587,7 +1587,7 @@ void CKeyEventSink::OnSyncConfig(const std::string& key, const std::vector<uint8
         uint8_t count = value[1];
 
         std::vector<std::pair<wchar_t, wchar_t>> pairs;
-        for (int i = 0; i < count && (2 + i * 4 + 4) <= value.size(); i++)
+        for (size_t i = 0; i < count && (2 + i * 4 + 4) <= value.size(); i++)
         {
             uint16_t left = *reinterpret_cast<const uint16_t*>(value.data() + 2 + i * 4);
             uint16_t right = *reinterpret_cast<const uint16_t*>(value.data() + 2 + i * 4 + 2);
